sidebar done, toc almost
