+++ qcodes/widgets/__init__.py
@@ -1,0 +1,3 @@
+from .display import display_auto
+from .plot_window_widget import PlotWindowWidget
+from .sidebar_widget import SidebarWidget
