+++ qcodes/widgets/sidebar_widget/main.css
@@ -1,185 +1,18 @@
-/*extracted from https://gist.github.com/magican/5574556*/
-
-
-/*background color for links when you mouse over it 
-This style is used on export, but can be overwritted in the livenotebook 
-by selecting a color in the nbextension-configurator*/
-
-
-
-#toc-wrapper li > span:hover {
-  background-color: #DAA520;
-}
-
-#toc-level0  a {
-  color: #333333;  /* default - color also loaded from notebook config */ 
-  text-decoration: none;
-}
-#navigate_menu li > span:hover {background-color: #f1f1f1}
-
-
-/* Move menus and tooolbar to the left, following @Kevin-McIsaac suggestion 
-div#menubar-container, div#header-container {
-width: auto;
-padding-left: 20px;
-}*/
-
-#navigate_menu {
-  /*display: block;*/
-  list-style-type: none;
-  max-width: 800px; 
-  min-width: 100px; 
-  width: 250px; 
-  overflow: auto;
-}
-
-
-#navigate_menu a {
-  list-style-type: none;
-  /*color: #333333; */  /* now specified in js */
-  text-decoration: none;
-}
-
-#navigate_menu li  {
-  padding-left: 0px;
-  clear: both;
-  list-style-type: none;
-}
-
-#navigate_menu-level0  {padding-left: 0px;}
-#navigate_menu-level0 ul  {padding-left: 0px;}
-
-
-.toc {
-  max-height: 500px;
-  padding: 0px;
-  overflow-y: auto;
-  font-weight: normal;
-  /*color: #333333;*/  /* now specified in js */
-  white-space: nowrap;
-  overflow-x: auto;
-} 
-
-.toc  ol.toc-item {
-    counter-reset: item;
-    list-style: none;
-    padding: 0.1em;
-  }
-
-.toc ol.toc-item li {
-    display: block;
-  }
-
-.toc ul.toc-item {
-    list-style-type: none;
-    padding: 0;
-    margin:  0;
-}
-
-.toc  ol.toc-item li:before {
-    font-size: 90%;
-    font-family: Georgia, Times New Roman, Times, serif;
-    counter-increment: item;
-    content: counters(item, ".")" ";
-  }
-
-
-.float-wrapper {
-  position: fixed !important;
-  top: 120px;
-  max-width:600px;
-  right: 20px;
-  border: thin solid rgba(0, 0, 0, 0.38);
-  border-radius: 5px;
-  padding:10px;
-  /*background-color: #fff;*/  /* now specified in js */
-  opacity: .8;
-  z-index: 100;
-  overflow: hidden;
-}
-
 .sidebar-wrapper {
     height: 100%;
-    left: 5px;
-    padding: 10px;
+    padding: 3px;
     position: fixed !important;
-    width: 212px;
-    max-width: 28%;
-    /*background-color: #fff;*/  /* now specified in js */
+    width: 250px;
+    max-width: 40%;
+    background-color: #FFFFFF;
     border-style: solid;
-    /*border-color: #eeeeee;*/  /* now specified in js */
+    border-color: #EEEEEE;
     opacity: .99;
     overflow: hidden;
 }
-
-
-.col-md-9 {
-  overflow:hidden;
-  margin-left: 14%;
-  width: 80%}
-
-#toc-wrapper.closed {
-  min-width: 100px;
-  width: auto;
-  transition: width;
+.sidebar-wrapper.left {
+    left: 0px;
 }
-#toc-wrapper:hover{
-  opacity: 1;
-}
-#toc-wrapper .header {
-  font-size: 18px;
-  font-weight: bold;
-}
-#toc-wrapper .hide-btn {
-  font-size: 14px;
-  font-family: monospace;
-}
-
-#toc-wrapper .reload-btn {
-  font-size: 14px;
-  font-family: monospace;
-}
-
-#toc-wrapper .number_sections-btn {
-  font-size: 14px;
-  font-family: monospace;
-}
-
-
-/* don't waste so much screen space... */
-
-/*
- .toc .toc-item .toc-item {
-  padding-left: 1em;
-}
-*/
-
-.toc-item li { margin:0; padding:0; color:black }
-.toc-item li > span { display:block }
-.toc-item li > span { padding-left:0em }
-.toc-item li li > span { padding-left:1em }
-.toc-item li li li > span { padding-left:2em }
-.toc-item li li li li > span { padding-left:3em }
-.toc-item li li li li li > span { padding-left:4em }
-.toc-item li li li li li li > span { padding-left:5em }
-
-.toc-item-num {
-    font-style: normal;
-}
-
-#toc-wrapper .toc-item-num {
-    font-style: normal;
-    font-family: Georgia, Times New Roman, Times, serif;
-    /*color: black;*/  /* now specified in js */
-}
-
-/*
-These colors are now specified in js, after reading the extension's config stored in system
-and updated using the nbextension-configurator
-.toc-item-highlight-select  {background-color: Gold} 
-.toc-item-highlight-execute  {background-color: red} 
-.toc-item-highlight-execute.toc-item-highlight-select   {background-color: Gold} */
-
-.toc-item .fa-fw:first-child {
-  cursor: pointer;
+.sidebar-wrapper.right {
+    right: 18px;
 }
