+++ qcodes/widgets/toc_widget/__init__.py
@@ -1,0 +1,1 @@
+from .toc_widget import TOCWidget
