+++ qcodes/widgets/toc_widget/toc_widget.py
@@ -1,0 +1,8 @@
+import atexit
+from IPython.display import display
+from ipywidgets.widgets import *
+from traitlets import Unicode, List, Bool
+
+
+class TOCWidget(DOMWidget):
+    pass
